requirements.txt
# Core ML and Data Processing
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
datasets>=2.14.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
pillow>=10.0.0

# Data Engineering and Streaming
kafka-python>=2.0.2
apache-beam[gcp]>=2.50.0
pyspark>=3.5.0
delta-spark>=2.4.0
feast>=0.34.0

# Vector Search and Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
qdrant-client>=1.6.0

# API and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
httpx>=0.25.0

# GenAI and LLM
langchain>=0.0.350
langchain-community>=0.0.10
openai>=1.3.0
anthropic>=0.7.0

# MLOps and Experiment Tracking
mlflow>=2.8.0
wandb>=0.16.0
dvc>=3.30.0

# Cloud and Infrastructure
boto3>=1.34.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.19.0
redis>=5.0.0

# Monitoring and Observability
prometheus-client>=0.19.0
grafana-client>=3.5.0
psutil>=5.9.0

# Development and Testing
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Visualization and Frontend Support
plotly>=5.17.0
streamlit>=1.28.0
bokeh>=3.3.0

# Production Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.0.1
python-multipart>=0.0.6

# Production Utilities
python-dotenv>=1.0.0
typer>=0.9.0
structlog>=23.2.0